Move data entry API handlers to a separate module
